chore: simplify example files
