+++ modules/gcloud/bigquery_test.go
@@ -1,0 +1,84 @@
+package gcloud_test
+
+import (
+	"context"
+	"errors"
+	"fmt"
+
+	"cloud.google.com/go/bigquery"
+	"google.golang.org/api/iterator"
+	"google.golang.org/api/option"
+	"google.golang.org/api/option/internaloption"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
+
+	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/modules/gcloud"
+)
+
+func ExampleRunBigQueryContainer() {
+	// runBigQueryContainer {
+	ctx := context.Background()
+
+	bigQueryContainer, err := gcloud.RunBigQueryContainer(
+		ctx,
+		testcontainers.WithImage("ghcr.io/goccy/bigquery-emulator:0.4.3"),
+		gcloud.WithProjectID("bigquery-project"),
+	)
+	if err != nil {
+		panic(err)
+	}
+
+	// Clean up the container
+	defer func() {
+		if err := bigQueryContainer.Terminate(ctx); err != nil {
+			panic(err)
+		}
+	}()
+	// }
+
+	// bigQueryClient {
+	projectID := bigQueryContainer.Settings.ProjectID
+
+	opts := []option.ClientOption{
+		option.WithEndpoint(bigQueryContainer.URI),
+		option.WithGRPCDialOption(grpc.WithTransportCredentials(insecure.NewCredentials())),
+		option.WithoutAuthentication(),
+		internaloption.SkipDialSettingsValidation(),
+	}
+
+	client, err := bigquery.NewClient(ctx, projectID, opts...)
+	if err != nil {
+		panic(err)
+	}
+	defer client.Close()
+	// }
+
+	createFnQuery := client.Query("CREATE FUNCTION testr(arr ARRAY<STRUCT<name STRING, val INT64>>) AS ((SELECT SUM(IF(elem.name = \"foo\",elem.val,null)) FROM UNNEST(arr) AS elem))")
+	_, err = createFnQuery.Read(ctx)
+	if err != nil {
+		panic(err)
+	}
+
+	selectQuery := client.Query("SELECT testr([STRUCT<name STRING, val INT64>(\"foo\", 10), STRUCT<name STRING, val INT64>(\"bar\", 40), STRUCT<name STRING, val INT64>(\"foo\", 20)])")
+	it, err := selectQuery.Read(ctx)
+	if err != nil {
+		panic(err)
+	}
+
+	var val []bigquery.Value
+	for {
+		err := it.Next(&val)
+		if errors.Is(err, iterator.Done) {
+			break
+		}
+		if err != nil {
+			panic(err)
+		}
+	}
+
+	fmt.Println(val)
+
+	// Output:
+	// [30]
+}
